faet: code duplicity
